add cluster field to controll default router accessibility
the `private` field of an OSD/ROSA `/openshift/cluster-1.yml` only affects the API accessibility. this PR introduces an additional cluster flag `ingressPrivate` to controll accessibility on the ingress router.

the default will be `false`, which implies that cluster currently having a private ingress controller, must be explicitely configured in app-interface before promoting the q-r change to the ocm-clusters that leverages the new field.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>