change-type multiple context extraction conditiona
dynamic context extraction for a changed file has a condition modifier that can be one of the following.

* added - which makes the change-type react when a context is added in the current MR
* removed - which makes the change-type react when a context is removed in the current MR
* None - which makes the change-type react when the context does not change

if a change type is supposed to react to multiple of those conditions, we need to be able to specify a list of conditions.

since making `when` would be a breaking change, we will deprecate it while introducing a replacement field `condition`.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>